Refactor processing into submodules
