+++ backend/src/processing/mod.rs
@@ -1,0 +1,4 @@
+pub mod ocr;
+pub mod parse;
+pub mod report;
+pub mod ai_client;
